Added and Took Away
So i added a bunch of stuff and moved other stuff around. I kept trying to get the logos to work but i couldn't. I tried to move them around and change their paths but I couldn't get the logo to appear. And i followed a couple you tube tutorials on how to make some extra functionality to add flare